+++ requirements.txt
@@ -1,14 +1,16 @@
-backports.statistics==0.1.0
-chardet==4.0.0
-colorama==0.4.4
-Cython==0.29.28
-dl==0.1.0
+backports.statistics~=0.1.0
+chardet~=4.0.0
+colorama~=0.4.4
+Cython~=0.29.28
+dl~=0.1.0
 docutils
-ipaddr==2.2.0
-keyring==23.0.1
-mypy_extensions==0.4.3
-numpy==1.19.5
-ordereddict==1.1
-Pillow==9.1.0
-Sphinx==4.5.0
-typing_extensions==4.2.0
+ipaddr~=2.2.0
+keyring~=23.0.1
+mypy_extensions~=0.4.3
+numpy~=1.19.5
+ordereddict~=1.1
+Pillow~=9.1.0
+Sphinx~=4.5.0
+typing_extensions~=4.2.0
+requests~=2.27.1
+Flask~=2.1.2
